Executed CP readonly - endpoint error handling

+++ rackcity/views/change_plan_views.py
@@ -13,6 +13,7 @@ from rackcity.models import (
 from rackcity.utils.change_planner_utils import (
     get_modifications_in_cp,
     asset_cp_has_conflicts,
+    get_cp_already_executed_response,
 )
 from rackcity.utils.errors_utils import (
     Status,
@@ -35,7 +36,6 @@ from rest_framework.decorators import permission_classes, api_view
 from rest_framework.parsers import JSONParser
 from rest_framework.permissions import IsAuthenticated
 
-
 @api_view(['POST'])
 @permission_classes([IsAuthenticated])
 def change_plan_resolve_conflict(request, id):
@@ -43,6 +43,9 @@ def change_plan_resolve_conflict(request, id):
     Resolve a merge conflict
     """
     (change_plan, response) = get_change_plan(id)
+    if response:
+        return response
+    response = get_cp_already_executed_response(change_plan)
     if response:
         return response
     data = JSONParser().parse(request)
@@ -103,6 +106,9 @@ def change_plan_remove_asset(request, id):
     Remove a single assetCP from a change plan
     """
     (change_plan, response) = get_change_plan(id)
+    if response:
+        return response
+    response = get_cp_already_executed_response(change_plan)
     if response:
         return response
     data = JSONParser().parse(request)
@@ -183,6 +189,9 @@ def change_plan_delete(request):
             },
             status=HTTPStatus.BAD_REQUEST
         )
+    response = get_cp_already_executed_response(existing_change_plan)
+    if response:
+        return response
     try:
         existing_change_plan.delete()
     except Exception as error:
@@ -236,6 +245,9 @@ def change_plan_modify(request):
             },
             status=HTTPStatus.BAD_REQUEST
         )
+    response = get_cp_already_executed_response(existing_change_plan)
+    if response:
+        return response
     for field in data.keys():
         value = data[field]
         setattr(existing_change_plan, field, value)
@@ -408,6 +420,7 @@ def change_plan_execute(request, id):
             },
             status=HTTPStatus.BAD_REQUEST,
         )
+
     assets_cp = AssetCP.objects.filter(change_plan=change_plan)
     for asset_cp in assets_cp:
         if asset_cp_has_conflicts(asset_cp):
